rm metadata file
