+++ funsor/torch/__init__.py
@@ -1,15 +1,11 @@
 # Copyright Contributors to the Pyro project.
 # SPDX-License-Identifier: Apache-2.0
 
-from collections import OrderedDict
-
 import torch
 from multipledispatch import dispatch
 
-from funsor.constant import Constant
 from funsor.tensor import tensor_to_funsor
 from funsor.terms import to_funsor
-from funsor.torch.provenance import ProvenanceTensor
 from funsor.util import quote
 
 from . import distributions as _
@@ -24,13 +20,6 @@ def _quote(x, indent, out):
     Work around PyTorch not supporting reproducible repr.
     """
     out.append((indent, "torch.tensor({}, dtype={})".format(repr(x.tolist()), x.dtype)))
-
-
-@to_funsor.register(ProvenanceTensor)
-def provenance_to_funsor(x, output=None, dim_to_name=None):
-    if isinstance(x, ProvenanceTensor):
-        ret = to_funsor(x._t, output=output, dim_to_name=dim_to_name)
-        return Constant(OrderedDict(x._provenance), ret)
 
 
 to_funsor.register(torch.Tensor)(tensor_to_funsor)
